train_step: bf16 gradient accumulation for the token/value embeddings
wte and value_embeds are the two biggest tensors in the model; fp32 grad
buffers doubled their backward scatter traffic and would double their
reduce_scatter bytes at world>1 (the A100 trace shows the embedding
backward among the largest single kernels). Their .grad32 buffers now
default to bf16 — matching the autograd baseline's numerics exactly
(bf16 params -> bf16 .grad there) — while everything else stays fp32.
The AdamW kernel upcasts the grad once so moment math stays fp32 and the
optimizer step remains bit-identical for equal-valued grads (eager
parity test still exact). Sharded comm buffers follow the grad dtype.
An explicit init_grad_buffers dtype still overrides everything (fp64
parity tier).

+++ tests/test_grad_parity.py
@@ -175,7 +175,9 @@ def test_grad_parity_fp32_sdpa_d12():
         idx, targets, cu = make_batch(32768, 8192, 2048, "cuda")
         loss_ref = model(idx, cu, targets)
         loss_ref.backward()
-        init_grad_buffers(model)
+        # explicit fp32 buffers: this tier is the precise guard, and under the
+        # fp32 COMPUTE_DTYPE patch the autograd reference grads are fp32 too
+        init_grad_buffers(model, dtype=torch.float32)
         loss = forward_backward(model, idx, targets, cu, loss_scale=1.0)
         assert abs(loss.item() - loss_ref.item()) / abs(loss_ref.item()) < 1e-6
         for name, p in model.named_parameters():
